Refactoring for AdvancedMenuNode
