+++ Resources/config/advanced.admin.xml
@@ -5,8 +5,8 @@
     xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
 
     <parameters>
-        <parameter key="cmf_menu.multilang.menu_admin_class">Symfony\Cmf\Bundle\MenuBundle\Admin\MultilangMenuAdmin</parameter>
-        <parameter key="cmf_menu.multilang.node_admin_class">Symfony\Cmf\Bundle\MenuBundle\Admin\MultilangMenuNodeAdmin</parameter>
+        <parameter key="cmf_menu.multilang.menu_admin_class">Symfony\Cmf\Bundle\MenuBundle\Admin\AdvancedMenuAdmin</parameter>
+        <parameter key="cmf_menu.multilang.node_admin_class">Symfony\Cmf\Bundle\MenuBundle\Admin\AdvancedMenuNodeAdmin</parameter>
     </parameters>
 
     <services>
