+++ Resources/config/doctrine/AdvancedMenu.phpcr.xml
@@ -2,14 +2,9 @@
     xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
 
     <document
-        name="Symfony\Cmf\Bundle\MenuBundle\Document\MultilangMenuNode"
+        name="Symfony\Cmf\Bundle\MenuBundle\Document\AdvancedMenu"
         translator="attribute"
         >
-
-        <locale name="locale"/>
-
-        <field name="label" type="string" translated="true"/>
-        <field name="uri" type="string" translated="true"/>
     </document>
 
 </doctrine-mapping>
