--- Resources/config/doctrine/MultilangMenu.phpcr.xml
@@ -1,12 +1,0 @@
-<doctrine-mapping xmlns="http://doctrine-project.org/schemas/phpcr/doctrine-mapping"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-
-    <document
-        name="Symfony\Cmf\Bundle\MenuBundle\Document\MultilangMenu"
-        translator="attribute"
-        >
-    </document>
-
-</doctrine-mapping>
-
-
