+++ Resources/translations/CmfMenuBundle.fr.yml
@@ -1,9 +1,9 @@
 dashboard:
     group_menu: Menu
     label_menu: Menus
-    label_multilang_menu: Menus multilingue
+    label_advanced_menu: Menus multilingue
     label_menu_node: Elément de menu
-    label_multilang_menu_node: Elément de menu multilingue
+    label_advanced_menu_node: Elément de menu multilingue
 
 breadcrumb:
     link_menu_list: Menus
@@ -14,14 +14,14 @@ breadcrumb:
     link_menu_node_create: Créer
     link_menu_node_edit: Éditer
     link_menu_node_delete: Supprimer
-    link_multilang_menu_list: Menus multilingue
-    link_multilang_menu_create: Créer
-    link_multilang_menu_edit: Editer
-    link_multilang_menu_delete: Supprimer
-    link_multilang_menu_node_list: Eléments de menu multilingues
-    link_multilang_menu_node_create: Créer
-    link_multilang_menu_node_edit: Éditer
-    link_multilang_menu_node_delete: Supprimer
+    link_advanced_menu_list: Menus multilingue
+    link_advanced_menu_create: Créer
+    link_advanced_menu_edit: Editer
+    link_advanced_menu_delete: Supprimer
+    link_advanced_menu_node_list: Eléments de menu multilingues
+    link_advanced_menu_node_create: Créer
+    link_advanced_menu_node_edit: Éditer
+    link_advanced_menu_node_delete: Supprimer
 
 list:
     label_id: Id
