+++ Tests/Functional/Document/MenuNodeTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Symfony\Cmf\Bundle\MenuBundle\Tests\Functional\Admin\Document;
+namespace Symfony\Cmf\Bundle\MenuBundle\Tests\Functional\Document;
 
 use Symfony\Cmf\Component\Testing\Functional\BaseTestCase;
 use Symfony\Cmf\Bundle\MenuBundle\Document\MenuNode;
@@ -34,18 +34,15 @@ class MenuNodeTest extends BaseTestCase
         return new MenuNode;
     }
 
-    public function testMenuNode()
+    protected function getMenuNodeData()
     {
-        $data = array(
+        return array(
             'name' => 'test-node',
             'label' => 'label_foobar',
             'uri' => 'http://www.example.com/foo',
             'route' => 'foo_route',
             'weakContent' => $this->weakContent,
             'hardContent' => $this->hardContent,
-            'publishable' => false,
-            'publishStartDate' => new \DateTime('2013-06-18'),
-            'publishEndDate' => new \DateTime('2013-06-18'),
             'attributes' => array(
                 'attr_foobar_1' => 'barfoo',
                 'attr_foobar_2' => 'barfoo',
@@ -74,6 +71,11 @@ class MenuNodeTest extends BaseTestCase
             'display' => false,
             'displayChildren' => false,
         );
+    }
+
+    public function testMenuNode()
+    {
+        $data = $this->getMenuNodeData();
 
         $menuNode = $this->getNewInstance();
         $refl = new \ReflectionClass($menuNode);
@@ -122,14 +124,5 @@ class MenuNodeTest extends BaseTestCase
 
         // test children
         $this->assertCount(1, $menuNode->getChildren());
-
-        // test publish start and end
-        $publishStartDate = $data['publishStartDate'];
-        $publishEndDate = $data['publishEndDate'];
-
-        $this->assertInstanceOf('\DateTime', $publishStartDate);
-        $this->assertInstanceOf('\DateTime', $publishEndDate);
-        $this->assertEquals($data['publishStartDate']->format('Y-m-d'), $publishStartDate->format('Y-m-d'));
-        $this->assertEquals($data['publishEndDate']->format('Y-m-d'), $publishEndDate->format('Y-m-d'));
     }
 }
